make things 32bit floats

+++ struct_opt/barrier_objective.py
@@ -4,9 +4,9 @@ import tensorflow as tf
 class Barrier():
 
     def __init__(self, max_constraint: float, barrier_width: float, barrier_size: float):
-        self.max_constraint = tf.constant(max_constraint, dtype=tf.float64)
-        self.barrier_width = tf.constant(barrier_width, dtype=tf.float64)
-        self.barrier_size = tf.constant(barrier_size, dtype=tf.float64)
+        self.max_constraint = tf.constant(max_constraint, dtype=tf.float32)
+        self.barrier_width = tf.constant(barrier_width, dtype=tf.float32)
+        self.barrier_size = tf.constant(barrier_size, dtype=tf.float32)
 
     def get_objective_function(self):
         @tf.function
